requirements.txt
streamlit
openai
openai-whisper
transformers
ffmpeg
pydub
moviepy